LayoutTests/inspector/dom-debugger/event-listener-breakpoints-custom-exact-case-sensitive-expected.txt

CONSOLE MESSAGE: BREAKPOINT ACTION LOG 1
CONSOLE MESSAGE: BREAKPOINT ACTION LOG 2
CONSOLE MESSAGE: BREAKPOINT ACTION LOG 3
CONSOLE MESSAGE: BREAKPOINT ACTION LOG 4
Tests for Event Listener breakpoints.


== Running test suite: DOMDebugger.Event.Listener
-- Running test case: DOMDebugger.Event.Listener.Custom.Exact.CaseSensitive.AddBreakpoint
Creating "custom" Event Breakpoint...
Adding "listener:custom" Event Breakpoint...
Firing "custom" on body...
PASS: Should pause before event handler is run.
CALL STACK:
0: [F] handleBody_custom
1: [F] trigger_custom
2: [P] Global Code
-- Running test teardown.

-- Running test case: DOMDebugger.Event.Listener.Custom.Exact.CaseSensitive.AddDisabledBreakpoint
Creating "custom" Event Breakpoint...
Adding "listener:custom" Event Breakpoint...
Disabling "listener:custom" Event Breakpoint...
Firing "custom" on body...
PASS: Should not pause for disabled breakpoint.
-- Running test teardown.

-- Running test case: DOMDebugger.Event.Listener.Custom.Exact.CaseSensitive.RemoveBreakpoint
Creating "custom" Event Breakpoint...
Adding "listener:custom" Event Breakpoint...
Removing "listener:custom" Event Breakpoint...
Firing "custom" on body...
PASS: Should not pause for removed breakpoint.
-- Running test teardown.

-- Running test case: DOMDebugger.Event.Listener.Custom.Exact.CaseSensitive.RemoveDisabledBreakpoint
Creating "custom" Event Breakpoint...
Adding "listener:custom" Event Breakpoint...
Disabling "listener:custom" Event Breakpoint...
Removing "listener:custom" Event Breakpoint...
Firing "custom" on body...
PASS: Should not pause for removed disabled breakpoint.
-- Running test teardown.

-- Running test case: DOMDebugger.Event.Listener.Custom.Exact.CaseSensitive.Options.Condition
Creating "custom" Event Breakpoint...
Adding "listener:custom" Event Breakpoint...

Setting condition to 'false'...

Triggering breakpoint...
PASS: Should not pause.

Triggering breakpoint...
PASS: Should not pause.

Setting condition to 'true'...

Triggering breakpoint...
PASS: Should pause.

Triggering breakpoint...
PASS: Should pause.

-- Running test case: DOMDebugger.Event.Listener.Custom.Exact.CaseSensitive.Options.Condition.ConsoleCommandLineAPI
Adding saved console value 'false'...
Creating "custom" Event Breakpoint...
Adding "listener:custom" Event Breakpoint...

Setting condition to saved console value...

Triggering breakpoint...
PASS: Should not pause.

Triggering breakpoint...
PASS: Should not pause.

Adding saved console value 'true'...
Setting condition to saved console value...

Triggering breakpoint...
PASS: Should pause.

Triggering breakpoint...
PASS: Should pause.

-- Running test case: DOMDebugger.Event.Listener.Custom.Exact.CaseSensitive.Options.IgnoreCount
Creating "custom" Event Breakpoint...
Adding "listener:custom" Event Breakpoint...

Setting ignoreCount to '2'...

Triggering breakpoint...
PASS: Should not pause.

Triggering breakpoint...
PASS: Should not pause.

Triggering breakpoint...
PASS: Should pause.

Triggering breakpoint...
PASS: Should pause.

-- Running test case: DOMDebugger.Event.Listener.Custom.Exact.CaseSensitive.Options.Action.Log
Creating "custom" Event Breakpoint...
Adding "listener:custom" Event Breakpoint...

Adding log action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing log action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing log action...
Enabling auto-continue...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should not pause.

Editing log action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should not pause.

-- Running test case: DOMDebugger.Event.Listener.Custom.Exact.CaseSensitive.Options.Actions.Evaluate
Creating "custom" Event Breakpoint...
Adding "listener:custom" Event Breakpoint...

Adding evaluate action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing evaluate action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing evaluate action...
Enabling auto-continue...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should not pause.

Editing evaluate action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should not pause.

-- Running test case: DOMDebugger.Event.Listener.Custom.Exact.CaseSensitive.Options.Actions.Evaluate.ConsoleCommandLineAPI
Adding saved console value '1'...
Creating "custom" Event Breakpoint...
Adding "listener:custom" Event Breakpoint...

Adding evaluate action using saved console value...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should pause.

Adding saved console value '2'...
Editing evaluate action using saved console value...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should pause.

Adding saved console value '3'...
Editing evaluate action using saved console value...
Enabling auto-continue...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should not pause.

Adding saved console value '4'...
Editing evaluate action using saved console value...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should not pause.
